requirements.txt
python-telegram-bot>=22.6
openpyxl==3.1.5
Pillow==10.4.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
supabase==2.4.5
postgrest-py==0.15.2
httpx==0.28.1
pydantic==2.5.0
